Issue #111: Final changes to tests and splitting code into mutliple classes
Final changes to tests and splitting code into mutliple classes
