+++ metricshub-engine/src/main/java/org/sentrysoftware/metricshub/engine/client/ClientsExecutor.java
@@ -23,6 +23,10 @@ package org.sentrysoftware.metricshub.engine.client;
 
 import io.opentelemetry.instrumentation.annotations.SpanAttribute;
 import io.opentelemetry.instrumentation.annotations.WithSpan;
+import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -48,11 +52,6 @@ import org.sentrysoftware.winrm.service.client.auth.AuthenticationEnum;
 import org.sentrysoftware.winrm.wql.WinRMWqlExecutor;
 import org.sentrysoftware.xflat.XFlat;
 import org.sentrysoftware.xflat.exceptions.XFlatException;
-
-import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeoutException;
 
 /**
  * The ClientsExecutor class provides utility methods for executing
