+++ metricshub-wbem-extension/src/main/java/org/sentrysoftware/metricshub/extension/wbem/WbemConfiguration.java
@@ -2,7 +2,7 @@ package org.sentrysoftware.metricshub.extension.wbem;
 
 /*-
  * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
- * MetricsHub Engine
+ * MetricsHub Wbem Extension
  * ჻჻჻჻჻჻
  * Copyright 2023 - 2024 Sentry Software
  * ჻჻჻჻჻჻
