+++ metricshub-wbem-extension/src/main/java/org/sentrysoftware/metricshub/extension/wbem/WbemSourceProcessor.java
@@ -1,0 +1,47 @@
+package org.sentrysoftware.metricshub.extension.wbem;
+
+/*-
+ * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
+ * MetricsHub Wbem Extension
+ * ჻჻჻჻჻჻
+ * Copyright 2023 - 2024 Sentry Software
+ * ჻჻჻჻჻჻
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
+ */
+
+import static org.sentrysoftware.metricshub.engine.common.helpers.MetricsHubConstants.AUTOMATIC_NAMESPACE;
+import static org.sentrysoftware.metricshub.engine.common.helpers.MetricsHubConstants.WMI_DEFAULT_NAMESPACE;
+
+import org.sentrysoftware.metricshub.engine.connector.model.monitor.task.source.WbemSource;
+import org.sentrysoftware.metricshub.engine.telemetry.TelemetryManager;
+
+public class WbemSourceProcessor {
+
+	/**
+	 * Get the namespace to use for the execution of the given {@link WbemSource} instance
+	 *
+	 * @param wbemSource {@link WbemSource} instance from which we want to extract the namespace. Expected "automatic", null or <em>any string</em>
+	 * @return {@link String} value
+	 */
+	String getNamespace(final WbemSource wbemSource, final TelemetryManager telemetryManager, final String connectorId) {
+		String namespace = wbemSource.getNamespace();
+		if (namespace == null) {
+			namespace = WMI_DEFAULT_NAMESPACE;
+		} else if (AUTOMATIC_NAMESPACE.equalsIgnoreCase(namespace)) {
+			namespace = telemetryManager.getHostProperties().getConnectorNamespace(connectorId).getAutomaticWbemNamespace();
+		}
+		return namespace;
+	}
+}
