+++ metricshub-wbem-extension/src/test/java/org/sentrysoftware/WbemExtensionTest.java
@@ -1,9 +1,28 @@
 package org.sentrysoftware;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.sentrysoftware.metricshub.engine.common.helpers.KnownMonitorType.HOST;
+import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.DOWN;
+import static org.sentrysoftware.metricshub.extension.wbem.WbemExtension.WBEM_TEST_QUERY;
+
 import com.fasterxml.jackson.databind.node.BooleanNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -25,41 +44,20 @@ import org.sentrysoftware.metricshub.extension.wbem.WbemExtension;
 import org.sentrysoftware.metricshub.extension.wbem.WbemRequestExecutor;
 import org.sentrysoftware.wbem.javax.wbem.WBEMException;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doCallRealMethod;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.doThrow;
-import static org.sentrysoftware.metricshub.engine.common.helpers.KnownMonitorType.HOST;
-import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.DOWN;
-import static org.sentrysoftware.metricshub.extension.wbem.WbemExtension.WBEM_TEST_QUERY;
-
 @ExtendWith(MockitoExtension.class)
 public class WbemExtensionTest {
+
 	private static final String CONNECTOR_ID = "connector";
 	private static final String HOST_NAME = "test-host" + UUID.randomUUID();
-	private static final String SUCCESS_RESPONSE = "Success";
 	private static final String BMC_KEY = "0x0102";
 	private static final String USERNAME = "testUser";
 	private static final String PASSWORD = "testPassword";
 	private static final String WBEM_CRITERION_TYPE = "wbem";
 	private static final String WBEM_TEST_NAMESPACE = "namespace";
-	private static final String RESULT = "result";
 
 	public static final List<List<String>> EXECUTE_WBEM_RESULT = Arrays.asList(
-			Arrays.asList("value1a", "value2a", "value3a"),
-			Arrays.asList("value1b", "value2b", "value3b")
+		Arrays.asList("value1a", "value2a", "value3a"),
+		Arrays.asList("value1b", "value2b", "value3b")
 	);
 
 	@Mock
@@ -74,96 +72,76 @@ public class WbemExtensionTest {
 		final Monitor hostMonitor = Monitor.builder().type(HOST.getKey()).isEndpoint(true).build();
 
 		final Map<String, Map<String, Monitor>> monitors = new HashMap<>(
-				Map.of(HOST.getKey(), Map.of(HOST_NAME, hostMonitor))
+			Map.of(HOST.getKey(), Map.of(HOST_NAME, hostMonitor))
 		);
 
 		final WbemConfiguration wbemConfiguration = WbemConfiguration
-				.builder()
-				.username(USERNAME)
-				.password(PASSWORD.toCharArray())
-				.timeout(120L)
-				.build();
+			.builder()
+			.username(USERNAME)
+			.password(PASSWORD.toCharArray())
+			.timeout(120L)
+			.build();
 
 		telemetryManager =
-				TelemetryManager
+			TelemetryManager
+				.builder()
+				.monitors(monitors)
+				.hostConfiguration(
+					HostConfiguration
 						.builder()
-						.monitors(monitors)
-						.hostConfiguration(
-								HostConfiguration
-										.builder()
-										.hostname(HOST_NAME)
-										.hostId(HOST_NAME)
-										.hostType(DeviceKind.OOB)
-										.configurations(Map.of(WbemConfiguration.class, wbemConfiguration))
-										.build()
-						)
-						.build();
+						.hostname(HOST_NAME)
+						.hostId(HOST_NAME)
+						.hostType(DeviceKind.OOB)
+						.configurations(Map.of(WbemConfiguration.class, wbemConfiguration))
+						.build()
+				)
+				.build();
 	}
 
 	@Test
 	void testCheckProtocolUp() throws ClientException {
-		// Create a telemetry manager using a WBEM HostConfiguration
-		initWbem();
-
 		{
+			initWbem();
+
 			// Mock a positive WBEM protocol health check response
 			doReturn(EXECUTE_WBEM_RESULT)
-					.when(wbemRequestExecutorMock)
-					.executeWbem(
-							anyString(),
-							any(WbemConfiguration.class),
-							anyString(),
-							anyString(),
-							any(TelemetryManager.class)
-					);
+				.when(wbemRequestExecutorMock)
+				.executeWbem(anyString(), any(WbemConfiguration.class), anyString(), anyString(), any(TelemetryManager.class));
 
 			// Start the WBEM Health Check strategy
 			wbemExtension.checkProtocol(telemetryManager);
 
 			assertEquals(
-					WbemExtension.UP,
-					telemetryManager.getEndpointHostMonitor().getMetric(WbemExtension.WBEM_UP_METRIC).getValue()
+				WbemExtension.UP,
+				telemetryManager.getEndpointHostMonitor().getMetric(WbemExtension.WBEM_UP_METRIC).getValue()
 			);
 		}
 
 		{
-			doThrow(new RuntimeException(new WBEMException("WBEM_INVALID_NAMESPACE")))
-					.when(wbemRequestExecutorMock)
-					.executeWbem(
-							anyString(),
-							any(WbemConfiguration.class),
-							eq(WBEM_TEST_QUERY),
-							eq(WBEM_TEST_NAMESPACE),
-							telemetryManager
-					);
+			initWbem();
 
-			doCallRealMethod().when(wbemRequestExecutorMock).isAcceptableException(any());
+			doThrow(new RuntimeException(new WBEMException(WBEMException.CIM_ERR_INVALID_NAMESPACE)))
+				.when(wbemRequestExecutorMock)
+				.executeWbem(anyString(), any(WbemConfiguration.class), anyString(), anyString(), any(TelemetryManager.class));
 
 			// Start the WBEM Health Check
 			wbemExtension.checkProtocol(telemetryManager);
 
 			assertEquals(
-					WbemExtension.UP,
-					telemetryManager.getEndpointHostMonitor().getMetric(WbemExtension.WBEM_UP_METRIC).getValue()
+				WbemExtension.UP,
+				telemetryManager.getEndpointHostMonitor().getMetric(WbemExtension.WBEM_UP_METRIC).getValue()
 			);
 		}
 	}
-
 
 	@Test
 	void testCheckWbemDownHealth() throws ClientException {
 		initWbem();
 
 		// Mock null WBEM protocol health check response
-		doReturn(EXECUTE_WBEM_RESULT)
-				.when(wbemRequestExecutorMock)
-				.executeWbem(
-						anyString(),
-						any(WbemConfiguration.class),
-						eq(WBEM_TEST_QUERY),
-						eq(WBEM_TEST_NAMESPACE),
-						eq(telemetryManager)
-				);
+		doReturn(null)
+			.when(wbemRequestExecutorMock)
+			.executeWbem(anyString(), any(WbemConfiguration.class), anyString(), anyString(), any(TelemetryManager.class));
 
 		// Start the WBEM Health Check strategy
 		wbemExtension.checkProtocol(telemetryManager);
@@ -175,12 +153,12 @@ public class WbemExtensionTest {
 	void testIsValidConfiguration() {
 		assertTrue(wbemExtension.isValidConfiguration(WbemConfiguration.builder().build()));
 		assertFalse(
-				wbemExtension.isValidConfiguration(
-						new IConfiguration() {
-							@Override
-							public void validateConfiguration(String resourceKey) throws InvalidConfigurationException {}
-						}
-				)
+			wbemExtension.isValidConfiguration(
+				new IConfiguration() {
+					@Override
+					public void validateConfiguration(String resourceKey) throws InvalidConfigurationException {}
+				}
+			)
 		);
 	}
 
@@ -217,23 +195,13 @@ public class WbemExtensionTest {
 		configuration.set("skipAuth", BooleanNode.valueOf(false));
 
 		assertEquals(
-				WbemConfiguration
-						.builder()
-						.username(USERNAME)
-						.password(PASSWORD.toCharArray())
-						.timeout(120L)
-						.build(),
-				wbemExtension.buildConfiguration(WBEM_CRITERION_TYPE, configuration, value -> value)
+			WbemConfiguration.builder().username(USERNAME).password(PASSWORD.toCharArray()).timeout(120L).build(),
+			wbemExtension.buildConfiguration(WBEM_CRITERION_TYPE, configuration, value -> value)
 		);
 
 		assertEquals(
-				WbemConfiguration
-						.builder()
-						.username(USERNAME)
-						.password(PASSWORD.toCharArray())
-						.timeout(120L)
-						.build(),
-				wbemExtension.buildConfiguration(WBEM_CRITERION_TYPE, configuration, null)
+			WbemConfiguration.builder().username(USERNAME).password(PASSWORD.toCharArray()).timeout(120L).build(),
+			wbemExtension.buildConfiguration(WBEM_CRITERION_TYPE, configuration, null)
 		);
 	}
 
@@ -243,43 +211,38 @@ public class WbemExtensionTest {
 
 		telemetryManager.getHostConfiguration().setConfigurations(Map.of());
 
-		final WbemCriterion wbemCriterion = WbemCriterion.builder().type(WBEM_CRITERION_TYPE).build();
+		final WbemCriterion wbemCriterion = WbemCriterion
+			.builder()
+			.type(WBEM_CRITERION_TYPE)
+			.query(WBEM_TEST_QUERY)
+			.build();
 
-		assertEquals(
-				CriterionTestResult.empty(),
-				wbemExtension.processCriterion(wbemCriterion, CONNECTOR_ID, telemetryManager)
-		);
+		assertFalse(wbemExtension.processCriterion(wbemCriterion, CONNECTOR_ID, telemetryManager).isSuccess());
 	}
 
 	@Test
 	void testProcessCriterionNullResultTest() throws ClientException {
 		initWbem();
 
-		final WbemCriterion wbemCriterion = WbemCriterion.builder().type(WBEM_CRITERION_TYPE).build();
+		final WbemCriterion wbemCriterion = WbemCriterion
+			.builder()
+			.type(WBEM_CRITERION_TYPE)
+			.query(WBEM_TEST_QUERY)
+			.build();
 
-		final WbemConfiguration configuration = (WbemConfiguration) telemetryManager
-				.getHostConfiguration()
-				.getConfigurations()
-				.get(WbemConfiguration.class);
+		doReturn(null)
+			.when(wbemRequestExecutorMock)
+			.executeWbem(anyString(), any(WbemConfiguration.class), anyString(), anyString(), any(TelemetryManager.class));
 
-		doReturn(null).when(wbemRequestExecutorMock).executeWbem(
-				anyString(),
-				configuration,
-				eq(WBEM_TEST_QUERY),
-				eq(WBEM_TEST_NAMESPACE),
-				telemetryManager);
-
-		final String message =
-				"Received <null> result after connecting to the WBEM BMC chip with the WBEM-over-LAN interface.";
 		final CriterionTestResult criterionTestResult = wbemExtension.processCriterion(
-				wbemCriterion,
-				CONNECTOR_ID,
-				telemetryManager
+			wbemCriterion,
+			CONNECTOR_ID,
+			telemetryManager
 		);
 
-		assertEquals(null, criterionTestResult.getResult());
+		assertEquals("No result.", criterionTestResult.getResult());
 		assertFalse(criterionTestResult.isSuccess());
-		assertEquals(message, criterionTestResult.getMessage());
+		assertTrue(criterionTestResult.getMessage().contains("No result."));
 		assertNull(criterionTestResult.getException());
 	}
 
@@ -287,25 +250,26 @@ public class WbemExtensionTest {
 	void testProcessCriterionOk() throws ClientException {
 		initWbem();
 
-		final WbemCriterion wbemCriterion = WbemCriterion.builder().type(WBEM_CRITERION_TYPE).build();
+		final WbemCriterion wbemCriterion = WbemCriterion
+			.builder()
+			.type(WBEM_CRITERION_TYPE)
+			.query(WBEM_TEST_QUERY)
+			.build();
 
-		final WbemConfiguration configuration = (WbemConfiguration) telemetryManager
-				.getHostConfiguration()
-				.getConfigurations()
-				.get(WbemConfiguration.class);
+		doReturn(EXECUTE_WBEM_RESULT)
+			.when(wbemRequestExecutorMock)
+			.executeWbem(anyString(), any(WbemConfiguration.class), anyString(), anyString(), any(TelemetryManager.class));
 
-		doReturn(EXECUTE_WBEM_RESULT).when(wbemRequestExecutorMock).executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
-
-		final String message = "Successfully connected to the WBEM BMC chip with the WBEM-over-LAN interface.";
+		final String message = "WbemCriterion test succeeded:";
 		final CriterionTestResult criterionTestResult = wbemExtension.processCriterion(
-				wbemCriterion,
-				CONNECTOR_ID,
-				telemetryManager
+			wbemCriterion,
+			CONNECTOR_ID,
+			telemetryManager
 		);
 
-		assertEquals(EXECUTE_WBEM_RESULT, criterionTestResult.getResult());
+		assertEquals(SourceTable.tableToCsv(EXECUTE_WBEM_RESULT, ";", false), criterionTestResult.getResult());
 		assertTrue(criterionTestResult.isSuccess());
-		assertEquals(message, criterionTestResult.getMessage());
+		assertTrue(criterionTestResult.getMessage().contains(message));
 		assertNull(criterionTestResult.getException());
 	}
 
@@ -314,18 +278,20 @@ public class WbemExtensionTest {
 		initWbem();
 
 		final WbemConfiguration configuration = (WbemConfiguration) telemetryManager
-				.getHostConfiguration()
-				.getConfigurations()
-				.get(WbemConfiguration.class);
+			.getHostConfiguration()
+			.getConfigurations()
+			.get(WbemConfiguration.class);
 
-		doReturn(EXECUTE_WBEM_RESULT).when(wbemRequestExecutorMock).executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
+		doReturn(EXECUTE_WBEM_RESULT)
+			.when(wbemRequestExecutorMock)
+			.executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
 		final SourceTable actual = wbemExtension.processSource(
-				WbemSource.builder().build(),
-				CONNECTOR_ID,
-				telemetryManager
+			WbemSource.builder().query(WBEM_TEST_QUERY).namespace(WBEM_TEST_NAMESPACE).build(),
+			CONNECTOR_ID,
+			telemetryManager
 		);
 
-		final SourceTable expected = SourceTable.builder().rawData(RESULT).build();
+		final SourceTable expected = SourceTable.builder().table(EXECUTE_WBEM_RESULT).build();
 		assertEquals(expected, actual);
 	}
 
@@ -334,22 +300,17 @@ public class WbemExtensionTest {
 		initWbem();
 
 		final WbemConfiguration configuration = (WbemConfiguration) telemetryManager
-				.getHostConfiguration()
-				.getConfigurations()
-				.get(WbemConfiguration.class);
+			.getHostConfiguration()
+			.getConfigurations()
+			.get(WbemConfiguration.class);
 
 		doThrow(new RuntimeException("exception"))
-				.when(wbemRequestExecutorMock)
-				.executeWbem(
-						anyString(),
-						configuration,
-						eq(WBEM_TEST_QUERY),
-						eq(WBEM_TEST_NAMESPACE),
-						eq(telemetryManager));
+			.when(wbemRequestExecutorMock)
+			.executeWbem(anyString(), eq(configuration), eq(WBEM_TEST_QUERY), eq(WBEM_TEST_NAMESPACE), eq(telemetryManager));
 		final SourceTable actual = wbemExtension.processSource(
-				WbemSource.builder().build(),
-				CONNECTOR_ID,
-				telemetryManager
+			WbemSource.builder().query(WBEM_TEST_QUERY).build(),
+			CONNECTOR_ID,
+			telemetryManager
 		);
 
 		final SourceTable expected = SourceTable.empty();
@@ -361,16 +322,18 @@ public class WbemExtensionTest {
 		initWbem();
 
 		final WbemConfiguration configuration = (WbemConfiguration) telemetryManager
-				.getHostConfiguration()
-				.getConfigurations()
-				.get(WbemConfiguration.class);
+			.getHostConfiguration()
+			.getConfigurations()
+			.get(WbemConfiguration.class);
 
 		{
-			doReturn(null).when(wbemRequestExecutorMock).executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
+			doReturn(null)
+				.when(wbemRequestExecutorMock)
+				.executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
 			final SourceTable actual = wbemExtension.processSource(
-					WbemSource.builder().build(),
-					CONNECTOR_ID,
-					telemetryManager
+				WbemSource.builder().query(WBEM_TEST_QUERY).build(),
+				CONNECTOR_ID,
+				telemetryManager
 			);
 
 			final SourceTable expected = SourceTable.empty();
@@ -378,14 +341,16 @@ public class WbemExtensionTest {
 		}
 
 		{
-			doReturn("").when(wbemRequestExecutorMock).executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
+			doReturn(Collections.emptyList())
+				.when(wbemRequestExecutorMock)
+				.executeWbem(HOST_NAME, configuration, WBEM_TEST_QUERY, WBEM_TEST_NAMESPACE, telemetryManager);
 			final SourceTable actual = wbemExtension.processSource(
-					WbemSource.builder().build(),
-					CONNECTOR_ID,
-					telemetryManager
+				WbemSource.builder().query(WBEM_TEST_QUERY).build(),
+				CONNECTOR_ID,
+				telemetryManager
 			);
 
-			final SourceTable expected = SourceTable.builder().rawData("").build();
+			final SourceTable expected = SourceTable.builder().rawData(null).build();
 			assertEquals(expected, actual);
 		}
 	}
@@ -397,8 +362,12 @@ public class WbemExtensionTest {
 		telemetryManager.getHostConfiguration().setConfigurations(Map.of());
 
 		assertEquals(
-				SourceTable.empty(),
-				wbemExtension.processSource(WbemSource.builder().build(), CONNECTOR_ID, telemetryManager)
+			SourceTable.empty(),
+			wbemExtension.processSource(
+				WbemSource.builder().query(WBEM_TEST_NAMESPACE).build(),
+				CONNECTOR_ID,
+				telemetryManager
+			)
 		);
 	}
 }
