+++ metricshub-wbem-extension/src/test/java/org/sentrysoftware/WbemRequestExecutorTest.java
@@ -1,14 +1,15 @@
 package org.sentrysoftware;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import org.junit.jupiter.api.Test;
 import org.sentrysoftware.metricshub.engine.strategy.utils.WqlDetectionHelper;
 import org.sentrysoftware.metricshub.extension.wbem.WbemRequestExecutor;
 import org.sentrysoftware.wbem.javax.wbem.WBEMException;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
 public class WbemRequestExecutorTest {
+
 	@Test
 	void testIsAcceptableException() {
 		assertFalse(WqlDetectionHelper.isAcceptableException(null));
@@ -22,7 +23,7 @@ public class WbemRequestExecutorTest {
 		assertTrue(WqlDetectionHelper.isAcceptableException(new WBEMException(WBEMException.CIM_ERR_INVALID_CLASS)));
 
 		assertTrue(
-				WqlDetectionHelper.isAcceptableException(new Exception(new WBEMException(WBEMException.CIM_ERR_NOT_FOUND)))
+			WqlDetectionHelper.isAcceptableException(new Exception(new WBEMException(WBEMException.CIM_ERR_NOT_FOUND)))
 		);
 	}
 }
